search sync worker now uses opensearch library for syncing
