+++ services/apps/search_sync_worker/src/bin/cleanup-all-members.ts
@@ -1,7 +1,5 @@
-import { DB_CONFIG, REDIS_CONFIG } from '../conf'
-import { InitService } from '../service/init.service'
-import { MemberSyncService } from '../service/member.sync.service'
-import { OpenSearchService } from '../service/opensearch.service'
+import { OpenSearchService, MemberSyncService, InitService } from '@crowd/opensearch'
+import { DB_CONFIG, OPENSEARCH_CONFIG, REDIS_CONFIG, SERVICE_CONFIG } from '../conf'
 import { DbStore, getDbConnection } from '@crowd/database'
 import { getServiceLogger } from '@crowd/logging'
 import { getRedisClient } from '@crowd/redis'
@@ -9,14 +7,14 @@ import { getRedisClient } from '@crowd/redis'
 const log = getServiceLogger()
 
 setImmediate(async () => {
-  const openSearchService = new OpenSearchService(log)
+  const openSearchService = new OpenSearchService(log, OPENSEARCH_CONFIG())
 
   const redis = await getRedisClient(REDIS_CONFIG(), true)
 
   const dbConnection = await getDbConnection(DB_CONFIG())
   const store = new DbStore(log, dbConnection)
 
-  const service = new MemberSyncService(redis, store, openSearchService, log)
+  const service = new MemberSyncService(redis, store, openSearchService, log, SERVICE_CONFIG())
 
   const pageSize = 100
   let results = await service.getAllIndexedTenantIds(pageSize)
