+++ services/apps/search_sync_worker/src/bin/cleanup-tenant-members.ts
@@ -1,6 +1,5 @@
-import { DB_CONFIG, REDIS_CONFIG } from '../conf'
-import { MemberSyncService } from '../service/member.sync.service'
-import { OpenSearchService } from '../service/opensearch.service'
+import { DB_CONFIG, OPENSEARCH_CONFIG, REDIS_CONFIG, SERVICE_CONFIG } from '../conf'
+import { OpenSearchService, MemberSyncService } from '@crowd/opensearch'
 import { DbStore, getDbConnection } from '@crowd/database'
 import { getServiceLogger } from '@crowd/logging'
 import { getRedisClient } from '@crowd/redis'
@@ -17,14 +16,14 @@ if (processArguments.length !== 1) {
 const tenantId = processArguments[0]
 
 setImmediate(async () => {
-  const openSearchService = new OpenSearchService(log)
+  const openSearchService = new OpenSearchService(log, OPENSEARCH_CONFIG())
 
   const redis = await getRedisClient(REDIS_CONFIG(), true)
 
   const dbConnection = await getDbConnection(DB_CONFIG())
   const store = new DbStore(log, dbConnection)
 
-  const service = new MemberSyncService(redis, store, openSearchService, log)
+  const service = new MemberSyncService(redis, store, openSearchService, log, SERVICE_CONFIG())
 
   await service.cleanupMemberIndex(tenantId)
