+++ services/apps/search_sync_worker/src/bin/create-alias.ts
@@ -1,5 +1,6 @@
-import { OpenSearchService } from '../service/opensearch.service'
+import { OpenSearchService } from '@crowd/opensearch'
 import { getServiceLogger } from '@crowd/logging'
+import { OPENSEARCH_CONFIG } from 'conf'
 
 const log = getServiceLogger()
 
@@ -14,7 +15,7 @@ const index = processArguments[0]
 const alias = processArguments[1]
 
 setImmediate(async () => {
-  const openSearchService = new OpenSearchService(log)
+  const openSearchService = new OpenSearchService(log, OPENSEARCH_CONFIG())
 
   await openSearchService.createAlias(index, alias)
   process.exit(0)
