+++ services/apps/search_sync_worker/src/bin/create-versioned-index.ts
@@ -1,6 +1,7 @@
-import { OpenSearchService } from '../service/opensearch.service'
+import { OpenSearchService } from '@crowd/opensearch'
 import { OpenSearchIndex } from '../types'
 import { getServiceLogger } from '@crowd/logging'
+import { OPENSEARCH_CONFIG } from 'conf'
 
 const log = getServiceLogger()
 
@@ -14,7 +15,7 @@ if (processArguments.length !== 1) {
 const index = processArguments[0]
 
 setImmediate(async () => {
-  const openSearchService = new OpenSearchService(log)
+  const openSearchService = new OpenSearchService(log, OPENSEARCH_CONFIG())
 
   await openSearchService.createIndexWithVersion(index as OpenSearchIndex)
   process.exit(0)
