+++ services/apps/search_sync_worker/src/bin/delete-index.ts
@@ -1,5 +1,6 @@
-import { OpenSearchService } from '../service/opensearch.service'
+import { OpenSearchService } from '@crowd/opensearch'
 import { getServiceLogger } from '@crowd/logging'
+import { OPENSEARCH_CONFIG } from 'conf'
 
 const log = getServiceLogger()
 
@@ -13,7 +14,7 @@ if (processArguments.length !== 1) {
 const index = processArguments[0]
 
 setImmediate(async () => {
-  const openSearchService = new OpenSearchService(log)
+  const openSearchService = new OpenSearchService(log, OPENSEARCH_CONFIG())
 
   await openSearchService.deleteIndex(index)
   process.exit(0)
