+++ services/apps/search_sync_worker/src/bin/get-doc-count.ts
@@ -1,5 +1,6 @@
-import { OpenSearchService } from '../service/opensearch.service'
+import { OpenSearchService } from '@crowd/opensearch'
 import { getServiceLogger } from '@crowd/logging'
+import { OPENSEARCH_CONFIG } from 'conf'
 
 const log = getServiceLogger()
 
@@ -13,7 +14,7 @@ if (processArguments.length !== 1) {
 const index = processArguments[0]
 
 setImmediate(async () => {
-  const openSearchService = new OpenSearchService(log)
+  const openSearchService = new OpenSearchService(log, OPENSEARCH_CONFIG())
 
   const result = await openSearchService.getDocumentCount(index)
   console.log(`Document count of ${index}:`, result)
