+++ services/apps/search_sync_worker/src/bin/sync-all-organizations.ts
@@ -1,9 +1,8 @@
-import { DB_CONFIG } from '../conf'
-import { OpenSearchService } from '../service/opensearch.service'
+import { OrganizationSyncService, OpenSearchService } from '@crowd/opensearch'
+import { DB_CONFIG, OPENSEARCH_CONFIG, SERVICE_CONFIG } from '../conf'
 import { DbStore, getDbConnection } from '@crowd/database'
 import { getServiceLogger } from '@crowd/logging'
 import { OrganizationRepository } from '../repo/organization.repo'
-import { OrganizationSyncService } from '../service/organization.sync.service'
 import { timeout } from '@crowd/common'
 
 const log = getServiceLogger()
@@ -11,7 +10,7 @@ const log = getServiceLogger()
 const MAX_CONCURRENT = 3
 
 setImmediate(async () => {
-  const openSearchService = new OpenSearchService(log)
+  const openSearchService = new OpenSearchService(log, OPENSEARCH_CONFIG())
 
   const dbConnection = await getDbConnection(DB_CONFIG())
   const store = new DbStore(log, dbConnection)
@@ -20,7 +19,7 @@ setImmediate(async () => {
 
   const tenantIds = await repo.getTenantIds()
 
-  const service = new OrganizationSyncService(store, openSearchService, log)
+  const service = new OrganizationSyncService(store, openSearchService, log, SERVICE_CONFIG())
 
   let current = 0
   for (let i = 0; i < tenantIds.length; i++) {
