+++ services/apps/search_sync_worker/src/bin/sync-tenant-activities.ts
@@ -1,6 +1,5 @@
-import { DB_CONFIG } from '../conf'
-import { ActivitySyncService } from '../service/activity.sync.service'
-import { OpenSearchService } from '../service/opensearch.service'
+import { ActivitySyncService, OpenSearchService } from '@crowd/opensearch'
+import { DB_CONFIG, OPENSEARCH_CONFIG } from '../conf'
 import { DbStore, getDbConnection } from '@crowd/database'
 import { getServiceLogger } from '@crowd/logging'
 
@@ -16,7 +15,7 @@ if (processArguments.length !== 1) {
 const tenantId = processArguments[0]
 
 setImmediate(async () => {
-  const openSearchService = new OpenSearchService(log)
+  const openSearchService = new OpenSearchService(log, OPENSEARCH_CONFIG())
 
   const dbConnection = await getDbConnection(DB_CONFIG())
   const store = new DbStore(log, dbConnection)
