+++ services/apps/search_sync_worker/src/repo/activity.repo.ts
@@ -1,38 +1,16 @@
 import { DbStore, RepositoryBase } from '@crowd/database'
 import { Logger } from '@crowd/logging'
-import { IDbActivitySyncData } from './activity.data'
+import { IDbActivityId } from './activity.data'
 
 export class ActivityRepository extends RepositoryBase<ActivityRepository> {
   constructor(dbStore: DbStore, parentLog: Logger) {
     super(dbStore, parentLog)
   }
 
-  public async getActivityData(activityIds: string[]): Promise<IDbActivitySyncData[]> {
+  public async checkActivitiesExist(activityIds: string[]): Promise<IDbActivityId[]> {
     const results = await this.db().any(
       `
-      select id,
-            "tenantId",
-            "segmentId",
-            type,
-            timestamp,
-            platform,
-            "isContribution",
-            score,
-            "sourceId",
-            "sourceParentId",
-            attributes,
-            channel,
-            body,
-            title,
-            url,
-            (sentiment -> 'sentiment')::int as sentiment,
-            "importHash",
-            "memberId",
-            "conversationId",
-            "parentId",
-            username,
-            "objectMemberId",
-            "objectMemberUsername"
+      select id
       from activities where id in ($(activityIds:csv)) and "deletedAt" is null
     `,
       {
@@ -41,130 +19,6 @@ export class ActivityRepository extends RepositoryBase<ActivityRepository> {
     )
 
     return results
-  }
-
-  public async checkActivitiesExist(tenantId: string, activityIds: string[]): Promise<string[]> {
-    const results = await this.db().any(
-      `
-      select id from activities where "tenantId" = $(tenantId) and id in ($(activityIds:csv)) and "deletedAt" is null
-    `,
-      {
-        tenantId,
-        activityIds,
-      },
-    )
-
-    return results.map((r) => r.id)
-  }
-
-  public async markSynced(activityIds: string[]): Promise<void> {
-    await this.db().none(
-      `update activities set "searchSyncedAt" = now() where id in ($(activityIds:csv))`,
-      {
-        activityIds,
-      },
-    )
-  }
-
-  public async getTenantActivitiesForSync(
-    tenantId: string,
-    perPage: number,
-    lastId?: string,
-  ): Promise<string[]> {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    let results: any[]
-
-    if (lastId) {
-      results = await this.db().any(
-        `
-      select id from activities 
-      where "tenantId" = $(tenantId) and "deletedAt" is null and id > $(lastId)
-      order by id
-      limit ${perPage};
-      `,
-        {
-          tenantId,
-          lastId,
-        },
-      )
-    } else {
-      results = await this.db().any(
-        `
-      select id from activities 
-      where "tenantId" = $(tenantId) and "deletedAt" is null
-      order by id
-      limit ${perPage};
-      `,
-        {
-          tenantId,
-        },
-      )
-    }
-
-    return results.map((r) => r.id)
-  }
-
-  public async getOrganizationActivitiesForSync(
-    organizationId: string,
-    perPage: number,
-    lastId?: string,
-  ): Promise<string[]> {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    let results: any[]
-
-    if (lastId) {
-      results = await this.db().any(
-        `
-      select id from activities 
-      where "organizationId" = $(organizationId) and "deletedAt" is null and id > $(lastId)
-      order by id
-      limit ${perPage};
-      `,
-        {
-          organizationId,
-          lastId,
-        },
-      )
-    } else {
-      results = await this.db().any(
-        `
-      select id from activities 
-      where "organizationId" = $(organizationId) and "deletedAt" is null
-      order by id
-      limit ${perPage};
-      `,
-        {
-          organizationId,
-        },
-      )
-    }
-
-    return results.map((r) => r.id)
-  }
-
-  public async getRemainingTenantActivitiesForSync(
-    tenantId: string,
-    page: number,
-    perPage: number,
-    cutoffDate: string,
-  ): Promise<string[]> {
-    const results = await this.db().any(
-      `
-      select id from activities 
-      where "tenantId" = $(tenantId) and "deletedAt" is null
-       and (
-        "searchSyncedAt" is null or
-        "searchSyncedAt" < $(cutoffDate)
-       )
-      limit ${perPage} offset ${(page - 1) * perPage};
-      `,
-      {
-        tenantId,
-        cutoffDate,
-      },
-    )
-
-    return results.map((r) => r.id)
   }
 
   public async getTenantIds(): Promise<string[]> {
