--- services/apps/search_sync_worker/src/repo/segment.data.ts
@@ -1,5 +1,0 @@
-export interface IDbSegmentInfo {
-  id: string
-  parentId: string
-  grandParentId: string
-}
