--- services/apps/search_sync_worker/src/service/member.sync.data.ts
@@ -1,4 +1,0 @@
-export interface IMemberSyncResult {
-  membersSynced: number
-  documentsIndexed: number
-}
