--- services/apps/search_sync_worker/src/service/opensearch.data.ts
@@ -1,15 +1,0 @@
-export interface IIndexRequest<T> {
-  id: string
-  body: T
-}
-
-export interface ISearchHit<T> {
-  _id: string
-
-  _source?: T
-}
-
-export interface IPagedSearchResponse<T, TAfter> {
-  data: T[]
-  afterKey?: TAfter
-}
