--- services/apps/search_sync_worker/src/service/organization.sync.data.ts
@@ -1,4 +1,0 @@
-export interface IOrganizationSyncResult {
-  organizationsSynced: number
-  documentsIndexed: number
-}
